refactor(trips): split monolithic service and controller into focused slices
trips.service.ts (497 lines) split into trips.service, trips-destinations.service,
and trips-groups.service. trips.controller.ts (297 lines) split into matching
controller files. Each file handles a single sub-domain: core trip lifecycle,
destination CRUD, or group-trip links. assertOrganizerRole made public so the
new services can delegate auth checks without circular deps.
