+++ apps/api/src/modules/trips/trips.controller.spec.ts
@@ -12,14 +12,6 @@ import type { CreateTripDto } from './dto/create-trip.dto';
 import type { UpdateTripDto } from './dto/update-trip.dto';
 import type { TransitionTripStatusDto } from './dto/transition-trip-status.dto';
 import type { TripResponseDto } from './dto/trip-response.dto';
-import type { CreateDestinationDto } from './dto/create-destination.dto';
-import type { UpdateDestinationDto } from './dto/update-destination.dto';
-import type { ReorderDestinationsDto } from './dto/reorder-destinations.dto';
-import type {
-  DestinationResponseDto,
-  DestinationWriteResponseDto,
-} from './dto/destination-response.dto';
-import type { TripGroupResponseDto } from './dto/trip-group-response.dto';
 import type { AuthenticatedUser } from '@/types/express';
 
 const mockUser: AuthenticatedUser = {
@@ -62,27 +54,6 @@ const mockResponse: TripResponseDto = {
   feedbackOpenUntil: null,
 };
 
-const mockDestResponse: DestinationResponseDto = {
-  id: 'dest-uuid',
-  tripId: 'trip-uuid',
-  position: 1,
-  countryCode: 'MX',
-  city: 'CANCUN',
-  label: null,
-  createdAt: '2026-01-01T00:00:00.000Z',
-};
-
-const mockDestWriteResponse: DestinationWriteResponseDto = {
-  ...mockDestResponse,
-  requiresConfirmation: false,
-};
-
-const mockGroupTripResponse: TripGroupResponseDto = {
-  tripId: 'trip-uuid',
-  groupId: 'group-uuid',
-  addedAt: '2026-01-01T00:00:00.000Z',
-};
-
 describe('TripsController', () => {
   let controller: TripsController;
   let mockCreateTrip: jest.Mock;
@@ -90,14 +61,6 @@ describe('TripsController', () => {
   let mockUpdateTrip: jest.Mock;
   let mockDeleteTrip: jest.Mock;
   let mockTransitionStatus: jest.Mock;
-  let mockListDestinations: jest.Mock;
-  let mockAddDestination: jest.Mock;
-  let mockUpdateDestination: jest.Mock;
-  let mockDeleteDestination: jest.Mock;
-  let mockReorderDestinations: jest.Mock;
-  let mockListTripGroups: jest.Mock;
-  let mockAddTripGroup: jest.Mock;
-  let mockRemoveTripGroup: jest.Mock;
 
   beforeEach(async () => {
     mockCreateTrip = jest.fn().mockResolvedValue(mockResponse);
@@ -105,14 +68,6 @@ describe('TripsController', () => {
     mockUpdateTrip = jest.fn().mockResolvedValue(mockResponse);
     mockDeleteTrip = jest.fn().mockResolvedValue(undefined);
     mockTransitionStatus = jest.fn().mockResolvedValue(mockResponse);
-    mockListDestinations = jest.fn().mockResolvedValue([mockDestResponse]);
-    mockAddDestination = jest.fn().mockResolvedValue(mockDestWriteResponse);
-    mockUpdateDestination = jest.fn().mockResolvedValue(mockDestWriteResponse);
-    mockDeleteDestination = jest.fn().mockResolvedValue(undefined);
-    mockReorderDestinations = jest.fn().mockResolvedValue([mockDestResponse]);
-    mockListTripGroups = jest.fn().mockResolvedValue([mockGroupTripResponse]);
-    mockAddTripGroup = jest.fn().mockResolvedValue(mockGroupTripResponse);
-    mockRemoveTripGroup = jest.fn().mockResolvedValue(undefined);
 
     const module: TestingModule = await Test.createTestingModule({
       controllers: [TripsController],
@@ -125,14 +80,6 @@ describe('TripsController', () => {
             updateTrip: mockUpdateTrip,
             deleteTrip: mockDeleteTrip,
             transitionStatus: mockTransitionStatus,
-            listDestinations: mockListDestinations,
-            addDestination: mockAddDestination,
-            updateDestination: mockUpdateDestination,
-            deleteDestination: mockDeleteDestination,
-            reorderDestinations: mockReorderDestinations,
-            listTripGroups: mockListTripGroups,
-            addTripGroup: mockAddTripGroup,
-            removeTripGroup: mockRemoveTripGroup,
           },
         },
       ],
@@ -190,67 +137,5 @@ describe('TripsController', () => {
 
     expect(mockTransitionStatus).toHaveBeenCalledWith(mockUser, 'trip-uuid', dto);
     expect(result).toBe(mockResponse);
-  });
-
-  it('listDestinations delegates to service', async () => {
-    const result = await controller.listDestinations('trip-uuid');
-
-    expect(mockListDestinations).toHaveBeenCalledWith('trip-uuid');
-    expect(result).toEqual([mockDestResponse]);
-  });
-
-  it('addDestination delegates to service', async () => {
-    const dto: CreateDestinationDto = { countryCode: 'MX', city: 'CANCUN' };
-
-    const result = await controller.addDestination(mockUser, 'trip-uuid', dto);
-
-    expect(mockAddDestination).toHaveBeenCalledWith(mockUser, 'trip-uuid', dto);
-    expect(result).toBe(mockDestWriteResponse);
-  });
-
-  it('updateDestination delegates to service', async () => {
-    const dto: UpdateDestinationDto = { city: 'TULUM' };
-
-    const result = await controller.updateDestination(mockUser, 'trip-uuid', 'dest-uuid', dto);
-
-    expect(mockUpdateDestination).toHaveBeenCalledWith(mockUser, 'trip-uuid', 'dest-uuid', dto);
-    expect(result).toBe(mockDestWriteResponse);
-  });
-
-  it('deleteDestination delegates to service', async () => {
-    await controller.deleteDestination(mockUser, 'trip-uuid', 'dest-uuid');
-
-    expect(mockDeleteDestination).toHaveBeenCalledWith(mockUser, 'trip-uuid', 'dest-uuid');
-  });
-
-  it('reorderDestinations delegates to service', async () => {
-    const dto: ReorderDestinationsDto = { destinationIds: ['dest-uuid'] };
-
-    const result = await controller.reorderDestinations(mockUser, 'trip-uuid', dto);
-
-    expect(mockReorderDestinations).toHaveBeenCalledWith(mockUser, 'trip-uuid', dto);
-    expect(result).toEqual([mockDestResponse]);
-  });
-
-  it('listTripGroups delegates to service', async () => {
-    const result = await controller.listTripGroups(mockUser, 'trip-uuid');
-
-    expect(mockListTripGroups).toHaveBeenCalledWith(mockUser, 'trip-uuid');
-    expect(result).toEqual([mockGroupTripResponse]);
-  });
-
-  it('addTripGroup delegates to service', async () => {
-    const dto = { groupId: 'group-uuid' };
-
-    const result = await controller.addTripGroup(mockUser, 'trip-uuid', dto);
-
-    expect(mockAddTripGroup).toHaveBeenCalledWith(mockUser, 'trip-uuid', dto);
-    expect(result).toBe(mockGroupTripResponse);
-  });
-
-  it('removeTripGroup delegates to service', async () => {
-    await controller.removeTripGroup(mockUser, 'trip-uuid', 'group-uuid');
-
-    expect(mockRemoveTripGroup).toHaveBeenCalledWith(mockUser, 'trip-uuid', 'group-uuid');
   });
 });
